Multiple sidebars, and better sidebar styling
The theme now defines sidebars for a variety of locations,
so a blog post sidebar, say, can contain different things
to the homepage sidebar, or the sidebar on an action guide.

We also include some styling for the "Helpful" plugin when
it is displayed in a sidebar.

And the primary and secondary column widths across the whole
site are now consistent, and work better at md and xl widths.

Fixes #6.