Optional “Show more” link for post list block
Generating a URL automatically from the query settings was too hard
(since the post list block allows filtering by multiple criteria) so
I’ve just allowed the editor to specify a URL to be used, and we leave
it up to them to decide if and when it should be shown. This feels like
the path of least surprise for editors.

Fixes #33.