Display post type (technically a category/term) in post lists
This should help people understand the different types of content
on the site, and is particularly useful on the search results page
and the new /groups directory page.

The category labels make more sense when they display in singular form
(ie: “Action guide” rather than “Action guides”). Annoyingly, WordPress
doesn’t give taxonomy terms a "singular_name" field (not to be confused
with `$taxonomy->labels->singular_name` which is the singular name for
the entire taxonomy, eg: “Category” or “Tag”).

To solve this, we add our own custom `singular_name` meta field on
the built-in `category` taxonomy. Admins will have to specify their
own singular names when setting up / editing categories. If a custom
singular name isn’t defined for a category, we fall back to the
(usually plural) category name.

We might want to extend this `singular_name` handling to our custom
`group-type` taxonomy too. But for now, all of the terms in that
taxonomy are already named in singular form, so it’s ok.