Vagrant provisioner: Use sudo when symlinking theme into place
We recently changed the provisioning script to make the web directory
owned by the web server process, so that WordPress can download updates
and plugins. But we forgot to update this one symlink line in the
provisioning script, so the symlink was failing to be created due to
a lack of permission to edit the web directory.

This line now uses sudo_symlink like all the other lines that
edit files in the web directory.

Thanks @wrightmartin for spotting the bug!

Fixes #61.