Move ids on headings to a child anchor, for better viewport jump
Fixes the situation where jumping to a heading (using a hash fragment
in the URL, eg: via the automatic Table of Contents) would result in
the heading itself being hidden behind any fixed-position banners at the
top of the window (eg: our customisable banners, or the WP admin bar).

This turned into a bit of a monster, as I ended up refactoring all the
code we were using to parse the innherHTML of the heading blocks. We now
use the built-in DOM library, rather than regular expressions, with a
few helper functions defined in the FMB/DOM namespace to handle the
transformation of the WordPress Heading block output into what we want.

Fixes #48.