Move Table of Contents generation into a Block
Tables of Contents are now no longer automatic. They must be requested
by adding a Block to the page.

This allows you to include a Table of Contents partway down a page,
for example, after an introductory paragraph.

We also now generate IDs for all heading blocks, so that Tables of
Contents can more easily be added, without the author having to manually
define an ID for each heading.

Fixes #63.